requirements/requirements.txt
apscheduler==3.6.3
awscli==1.20.38
black==21.8b0
boto3
fastapi>=0.68.1
mlflow>=1.18.0
pickle5==0.0.11
pre-commit==2.15.0
prometheus_client==0.7.1
pytest>=5.4.3
pytz==2019.3
pyyaml==5.3.1
requests>=2.22.0
s3fs==0.4.2
simplejson>=3.17.0
statsmodels>=0.12.2
urllib3>=1.25.10
uvicorn>= 0.15.0
